+++ components/script_servoparser/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "script_servoparser"
+version = "0.0.1"
+authors = ["The Servo Project Developers"]
+license = "MPL-2.0"
+publish = false
+
+build = "build.rs"
+
+[lib]
+name = "script_servoparser"
+path = "lib.rs"
+
+[dependencies]
+script = {path = "../script"}
+html5ever = "0.22"
+dom_struct = {path = "../dom_struct"}
+embedder_traits = {path = "../embedder_traits"}
+hyper = "0.10"
+msg = {path = "../msg"}
+net_traits = {path = "../net_traits"}
+profile_traits = {path = "../profile_traits"}
+script_traits = {path = "../script_traits"}
+servo_config = {path = "../config"}
+servo_url = {path = "../url"}
+style = {path = "../style", features = ["servo"]}
+mozjs = { version = "0.6", features = ["promises"]}
